pyproject.toml
[tool.poetry]
name = "thegadget"
version = "0.4.0"
description = "A strategic simulation, written in Python."
authors = ["Michael John <michael.john@gmx.at>"]
license = "MIT"
readme = "README.md"
include = [
    "thegadget/locales/**/LC_MESSAGES/*.mo",
    "thegadget/resources/database/*",
    "thegadget/resources/images/*",
    "thegadget/resources/maps/*",
    "thegadget/resources/music/*",
]
exclude = [
    "thegadget/resources/unused/*",
]

[tool.poetry.scripts]
thegadget = 'thegadget.__main__:main'

[tool.poetry.dependencies]
python = "^3.8"
# removed due to incompatibilities: pygame = "^2.4.0"
python-gettext = "^5.0"
pygame-ce = "^2.2.1"
pygame-gui = "^0.6.9"
cairosvg = "^2.7.0"
# removed due to incompatibilities: pygame-menu = "^4.4.3"
pygame-menu-ce = "^4.4.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
beautifulsoup4 = "^4.12.2"
python-gettext = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
homepage = "https://github.com/amstelchen/TheGadget"
repository = "https://github.com/amstelchen/TheGadget"
documentation = "https://github.com/amstelchen/TheGadget"

[tool.ruff]
line-length = 120
ignore = [
    "F401",
    "F403",
    "F405",
    "F841",
    "E501",
    "E701",
    "E702",
]

[tool.flake8]
ignore = [
    "F401",
    "F403",
    "F405",
    "F841",
    "E501",
    "E701",
    "E702",
]